split large and growing test file to three
